add smoke tests
