+++ src/app/decks/Decks.css
@@ -31,7 +31,6 @@
   hyphens: auto;
   justify-content: center;
   text-align: center;
-  height: 200px;
 }
 
 .deck-card:hover {
