+++ src/app/home/Home.js
@@ -1,14 +1,15 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
 import pluralize from "pluralize";
-import { Grid, Header, Icon, Label, Popup, Progress, Segment } from "semantic-ui-react";
-import { BarChart, Bar, XAxis, YAxis, Rectangle, ResponsiveContainer } from "recharts";
+import { Grid, Header, Icon, Label, Popup, Segment } from "semantic-ui-react";
 
 import * as api from "./homeActions";
 
+import ActivityOverview from "./ActivityOverview";
+import RecentActivity from "./RecentActivity";
+
 class Home extends Component {
   state = {
-    study: {},
     learn: {},
     review: {},
   };
@@ -20,8 +21,8 @@ class Home extends Component {
   fetchStudyTypes = () => {
     api.fetchStudyTypes().then(
       ({ data }) => {
-        const { study, learn, review } = data;
-        this.setState(() => ({ study, learn, review }));
+        const { learn, review } = data;
+        this.setState(() => ({ learn, review }));
       },
       error => {
         console.log("error", error);
@@ -43,47 +44,12 @@ class Home extends Component {
   render() {
     const { learn, review } = this.state;
 
-    // TODO: FAKE DATA. Replace with real values
-    const data = [
-      { name: "Mon", value: 100 },
-      { name: "Tue", value: 124 },
-      { name: "Wed", value: 30 },
-      { name: "Thu", value: 65 },
-      { name: "Fri", value: 40 },
-      { name: "Sat", value: 50 },
-      { name: "Sun", value: 200 },
-    ];
-
     return (
       <div className="study-page">
         <div className="container">
           <div className="row">
             <div className="col-md-10 offset-md-1 col-lg-8 offset-lg-2">
-              <Segment className="mb-3">
-                <Grid columns="3" textAlign="center">
-                  <Grid.Column>
-                    <span className="text-secondary">
-                      <Icon name="folder open outline" />
-                      <strong className="text-dark">10</strong> decks
-                    </span>
-                  </Grid.Column>
-                  <Grid.Column>
-                    <span className="text-secondary">
-                      <Icon name="file outline" />
-                      <strong className="text-dark">124</strong> cards
-                    </span>
-                  </Grid.Column>
-                  <Grid.Column>
-                    <span className="text-secondary">
-                      <Icon name="repeat" />
-                      <strong className="text-dark">100</strong> reviews
-                    </span>
-                  </Grid.Column>
-                </Grid>
-                <Progress attached="bottom" total={120} value={120} color="grey" />
-                <Progress attached="bottom" total={120} value={70} color="orange" />
-                <Progress attached="bottom" total={120} value={50} color="teal" />
-              </Segment>
+              <ActivityOverview />
               <div className="d-flex justify-content-between align-items-end mb-2">
                 <Header as="h3" className="m-0">
                   Choose your study type:
@@ -113,9 +79,11 @@ class Home extends Component {
                       introducing you to new material and expanding your knowledge base. This is
                       good if you are trying to learn a lot of material in a short time.
                     </p>
-                    <Label attached="top right" color="teal">
-                      {pluralize("card", learn.total, true)}
-                    </Label>
+                    {learn && (
+                      <Label attached="top right" color="teal">
+                        {pluralize("card", learn.total, true)}
+                      </Label>
+                    )}
                   </Segment>
                 </Grid.Column>
                 <Grid.Column>
@@ -135,41 +103,15 @@ class Home extends Component {
                       learned. These items are due to be reviewed again and will not contain items
                       you have just learned.
                     </p>
-                    <Label attached="top right" color="orange">
-                      {pluralize("card", review.total, true)}
-                    </Label>
+                    {review && (
+                      <Label attached="top right" color="orange">
+                        {pluralize("card", review.total, true)}
+                      </Label>
+                    )}
                   </Segment>
                 </Grid.Column>
               </Grid>
-              <Segment className="mt-4">
-                <Header>Recent Activity</Header>
-                <div className="border bg-light rounded p-2">
-                  <ResponsiveContainer height={200} width="100%">
-                    <BarChart data={data} padding={{ top: 10, bottom: 0, left: 0, right: 0 }}>
-                      <Bar
-                        dataKey="value"
-                        fill="#00b5ad"
-                        barSize={40}
-                        shape={<Rectangle radius={3} />}
-                      />
-                      <XAxis
-                        dataKey="name"
-                        axisLine={false}
-                        tickLine={false}
-                        height={15}
-                        fontSize="12px"
-                      />
-                      <YAxis
-                        dataKey="value"
-                        axisLine={false}
-                        tickLine={false}
-                        width={30}
-                        fontSize="12px"
-                      />
-                    </BarChart>
-                  </ResponsiveContainer>
-                </div>
-              </Segment>
+              <RecentActivity />
             </div>
           </div>
         </div>
