+++ src/app/home/RecentActivity.test.js
@@ -1,0 +1,8 @@
+import React from "react";
+import { shallow } from "enzyme";
+import RecentActivity from "./RecentActivity";
+
+it("renders without crashing", () => {
+  const wrapper = shallow(<RecentActivity />);
+  expect(wrapper).toHaveLength(1);
+});
